requirements.txt
python-telegram-bot==20.6
yfinance
requests
schedule
numpy
pandas
scikit-learn
aiohttp
nest_asyncio
openai
fpdf2
pandas
numpy==1.24.4
yfinance
requests
python-telegram-bot==20.7
schedule
scikit-learn
nest_asyncio
ta
